Add files via upload (#1)
* Add files via upload

Created the ability to save the output search data into a .csv file. The user must now verify if they wish to save, and if so, enter the save to file path. Also, added some user requirements to the Readme.

* Update README.md

Add changes to the README to explain functionality. Links, code, and images still needed, in addition to remaining sections.

* Update README.md

Removed filler text from template where applicable. Installation and Usage sections still need to be re-worked.

* Update README.md

Format corrected for list items.